pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ml-sgs-turbulence"
version = "1.0.0"
description = "Physics-Guided Multi-Task Learning for SGS Turbulence Parameterisation"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Sambit Kumar Panda",   email = "s.k.panda@example.ac.uk" },
    { name = "Todd Jones" },
    { name = "Muhammad Shahzad" },
    { name = "Bryan Lawrence" },
    { name = "Anna-Louise Ellis" },
]
keywords = [
    "turbulence", "subgrid-scale", "large-eddy simulation",
    "multi-task learning", "physics-guided", "neural network",
    "atmospheric modelling", "MONC",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pandas>=2.0",
    "h5py>=3.9",
    "torch>=2.0",
    "xarray>=2023.1",
    "netCDF4>=1.6",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.4",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=1.3",
    "myst-parser>=2.0",
    "nbsphinx>=0.9",
]

[project.scripts]
ml-sgs-compare   = "scripts.compare_models:main"
ml-sgs-process   = "ml_sgs.data.processor:main"

[project.urls]
Homepage   = "https://github.com/your-org/ml-sgs-turbulence"
Repository = "https://github.com/your-org/ml-sgs-turbulence"
Issues     = "https://github.com/your-org/ml-sgs-turbulence/issues"
Paper      = "https://doi.org/10.XXXX/XXXXXXX"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

# ---------------------------------------------------------------------------
# Ruff – linting & formatting
# ---------------------------------------------------------------------------
[tool.ruff]
line-length    = 100
target-version = "py310"
src            = ["src", "scripts", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "N806",   # variable in function should be lowercase (scientific code)
    "B008",   # do not perform function calls in default arguments
]

[tool.ruff.lint.per-file-ignores]
"tests/*"    = ["S101"]   # assert statements OK in tests
"scripts/*"  = ["T201"]   # print OK in entry-point scripts

# ---------------------------------------------------------------------------
# Mypy – static type checking
# ---------------------------------------------------------------------------
[tool.mypy]
python_version    = "3.10"
ignore_missing_imports = true
warn_unused_ignores    = true
warn_return_any        = false
files = ["src/ml_sgs"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths     = ["tests"]
addopts       = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source   = ["src/ml_sgs"]
omit     = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
